requirements.txt
httpx>=0.27.0
typer>=0.12.0
rich>=13.7.0
loguru>=0.7.2
python-dotenv>=1.0.1
